Update patch.sh
check if original files exist

+++ patch.sh
@@ -2,10 +2,18 @@
 
 parent="$(dirname "$(pwd)")"
 
-mv "${parent}/models/research/object_detection/builders/preprocessor_builder.py" "${parent}/models/research/object_detection/builders/preprocessor_builder.py.orig"
-mv "${parent}/models/research/object_detection/core/preprocessor.py"             "${parent}/models/research/object_detection/core/preprocessor.py.orig"
-mv "${parent}/models/research/object_detection/core/preprocessor_cache.py"       "${parent}/models/research/object_detection/core/preprocessor_cache.py.orig"
-mv "${parent}/models/research/object_detection/protos/preprocessor.proto"        "${parent}/models/research/object_detection/protos/preprocessor.proto.orig"
+if [ ! -f "${parent}/models/research/object_detection/builders/preprocessor_builder.py.orig" ]; then
+    mv "${parent}/models/research/object_detection/builders/preprocessor_builder.py" "${parent}/models/research/object_detection/builders/preprocessor_builder.py.orig"
+fi
+if [ ! -f "${parent}/models/research/object_detection/core/preprocessor.py.orig" ]; then
+    mv "${parent}/models/research/object_detection/core/preprocessor.py"             "${parent}/models/research/object_detection/core/preprocessor.py.orig"
+fi
+if [ ! -f "${parent}/models/research/object_detection/core/preprocessor_cache.py.orig" ]; then
+    mv "${parent}/models/research/object_detection/core/preprocessor_cache.py"       "${parent}/models/research/object_detection/core/preprocessor_cache.py.orig"
+fi
+if [ ! -f "${parent}/models/research/object_detection/protos/preprocessor.proto.orig" ]; then
+    mv "${parent}/models/research/object_detection/protos/preprocessor.proto"        "${parent}/models/research/object_detection/protos/preprocessor.proto.orig"
+fi
 
 cp ./object_detection/builders/preprocessor_builder.py "${parent}/models/research/object_detection/builders/"
 cp ./object_detection/core/preprocessor.py "${parent}/models/research/object_detection/core/"
